Added some new launcher features.
Ability to reset the patch version.
Ability to optimize the client by selectively "disabling" certain files, typically .nif files, in order to mitigate in-game lag. The list of files can be modified and updated as desired.
Ability to delete specific files. This is useful, especially when you accidentally patch a server-side file, such as 'IteminfoServer.shn.' Additionally, you can also add and update files to the list.
Ability to perform integrity checks. By using this, a player can better pinpoint the source of a possible "Client Manipulation" error.

Refactored a little bit and updated the documentation as well.